fix(supabase): corrige salvamento de registros e aplica novo tema azul
- Remove clear_on_submit e iniciar_pagina('app') que invalidavam as chaves dos widgets no submit (valores perdidos -> registro nunca inserido)

- Redesenha UI com paleta azul minimalista profissional (Inter, cards suaves, sidebar azul-marinho)

- Sincroniza tema do config.toml e ignora logs

+++ APIs/SupaBase/sql_criacao_tabelas.sql
@@ -73,9 +73,9 @@ ON CONFLICT ("NomeAreaResponsavel") DO NOTHING;
 CREATE TABLE "fDados" (
     "Id" BIGSERIAL PRIMARY KEY,
     "NumeroPendencia" INTEGER NOT NULL UNIQUE,
-    "IdUsuario" BIGINT NOT NULL REFERENCES "dUsuarios"("IdUsuario"),
-    "IdJustificativa" BIGINT NOT NULL REFERENCES "dJustificativas"("IdJustificativa"),
-    "IdAreaResponsavel" BIGINT NOT NULL REFERENCES "dAreasResponsaveis"("IdAreaResponsavel"),
+    "IdUsuario" INTEGER NOT NULL REFERENCES "dUsuarios"("IdUsuario"),
+    "IdJustificativa" INTEGER NOT NULL REFERENCES "dJustificativas"("IdJustificativa"),
+    "IdAreaResponsavel" INTEGER NOT NULL REFERENCES "dAreasResponsaveis"("IdAreaResponsavel"),
     "DataHora" TIMESTAMPTZ NOT NULL DEFAULT NOW()
 );
